GameSceneのScene遷移を管理するGameMaseterObjectの追加
