+++ Assets/Scripts/Scene/GameMaster.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54e6a291c3a7dd343b42bb3a6ff37d60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
